[ fix ] storage/ModelMapper 클래스 생성 및 기존 분리되있던 mapper 클래스 삭제
