--- movie-infrastructures/storage/src/main/java/com/movie/storage/movie/mapper/ScheduleMapper.java
@@ -1,16 +1,0 @@
-package com.movie.storage.movie.mapper;
-
-import com.movie.domain.movie.domain.Schedule;
-import com.movie.storage.movie.entity.ScheduleEntity;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class ScheduleMapper {
-
-    public static Schedule from(ScheduleEntity scheduleEntity) {
-        return null;
-    }
-
-}
